Replace TradePair in TradeCardGP (Gnosis)
